test: reorganizer test folder structure
Now emulates lib folder structure more accurately.
