+++ test/bin/test-citgm-all.js
@@ -3,9 +3,9 @@
 var path = require('path');
 
 var test = require('tap').test;
-var spawn = require('../lib/spawn');
+var spawn = require('../../lib/spawn');
 
-var citgmAllPath = path.resolve(__dirname, '..', 'bin', 'citgm-all');
+var citgmAllPath = path.resolve(__dirname, '..', '..', 'bin', 'citgm-all');
 
 test('citgm-all: /w markdown', function (t) {
   t.plan(1);
