+++ test/npm/test-npm-script-run.js
@@ -6,11 +6,13 @@ var mkdirp = require('mkdirp');
 var rimraf = require('rimraf');
 var rewire = require('rewire');
 
-var run = rewire('../lib/npm/script/run');
+var run = rewire('../../lib/npm/script/run');
 
-var passingScript = path.join(__dirname, 'fixtures', 'example-test-script-passing.sh');
-var failingScript = path.join(__dirname, 'fixtures', 'example-test-script-failing.sh');
-var badPath = path.join(__dirname, 'fixtures', 'example-test-script-does-not-exist');
+var fixtures = path.join(__dirname, '..', 'fixtures');
+
+var passingScript = path.join(fixtures, 'example-test-script-passing.sh');
+var failingScript = path.join(fixtures, 'example-test-script-failing.sh');
+var badPath = path.join(fixtures, 'example-test-script-does-not-exist');
 
 var sandbox = path.join(os.tmpdir(), 'citgm-' + Date.now() + 'run-test');
